Возобновление работы над программой (Внесены некоторые изменения)

+++ ComicLib/Pages/RegisterPage.xaml.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
+using System.Text.RegularExpressions;
 using System.Threading.Tasks;
 using System.Windows;
 using System.Windows.Controls;
@@ -12,6 +13,7 @@ using System.Windows.Media;
 using System.Windows.Media.Imaging;
 using System.Windows.Navigation;
 using System.Windows.Shapes;
+using ComicLib.ClassHelpers;
 
 namespace ComicLib.Pages
 {
@@ -23,15 +25,45 @@ namespace ComicLib.Pages
         public RegisterPage()
         {
             InitializeComponent();
+
+            if (NavigationService != null)
+            {
+                if (NavigationService.CanGoBack)
+                {
+                    NavigationService.RemoveBackEntry();
+                }
+            }
         }
 
         private void RegClick(object sender, RoutedEventArgs e)
         {
-            if (NavigationService.CanGoBack)
+            if (login.Text.Length == 0 || email.Text.Length == 0 || pas.Text.Length == 0)
             {
-                NavigationService.RemoveBackEntry();
+                errorMes.Text = "Все поля должны быть заполнены!";
             }
-            NavigationService.Navigate(new ViewPort());
+            else if (DBHelper.ValidateEmail(email.Text, out string message) && PasswordHelper.ValidatePassword(pas.Text, out message))
+            {
+                User user = new User
+                {
+                    Name = login.Text,
+                    Email = email.Text,
+                    Password = pas.Text
+                };
+                DBHelper.DBContext.User.Add(user);
+                DBHelper.DBContext.SaveChanges();
+
+                DBHelper.CurrentUser = DBHelper.DBContext.User.Where(u => u.Name == login.Text && u.Password == pas.Text).FirstOrDefault();
+
+                if (NavigationService.CanGoBack)
+                {
+                    NavigationService.RemoveBackEntry();
+                }
+                NavigationService.Navigate(new ViewPort());
+            }
+            else
+            {
+                errorMes.Text = message;
+            }
         }
 
         private void LoginPageClick(object sender, RoutedEventArgs e)
@@ -41,6 +73,35 @@ namespace ComicLib.Pages
                 NavigationService.RemoveBackEntry();
             }
             NavigationService.Navigate(new LoginPage());
+        }
+
+        private void TextBoxPreviewKeyDown(object sender, KeyEventArgs e)
+        {
+            e.Handled = e.Key == Key.Space;
+        }
+
+        private void LoginPreviewTextInput(object sender, TextCompositionEventArgs e)
+        {
+            if (!Regex.Match(e.Text, @"[0-9a-zA-Zа-яА-Я]").Success)
+            {
+                e.Handled = true;
+            }
+        }
+
+        private void PasswordPreviewTextInput(object sender, TextCompositionEventArgs e)
+        {
+            if (!Regex.Match(e.Text, @"[0-9a-zA-Z!@#$%^&*()_+=\[{\]};:<>|./?,-]").Success)
+            {
+                e.Handled = true;
+            }
+        }
+
+        private void EmailPreviewTextInput(object sender, TextCompositionEventArgs e)
+        {
+            if (!Regex.Match(e.Text, @"[a-z0-9@!#$%&.'*+/=?^_`{|}~-]").Success)
+            {
+                e.Handled = true;
+            }
         }
     }
 }
